refactor: delete shared/api/supabaseReads — fully split into feature/api/
